Some updates from review
